+++ src/providers/registry/AIProviderRegistry.ts
@@ -1,0 +1,8 @@
+import { ProviderRegistry } from './ProviderRegistry';
+
+export interface IAIProviderContract {
+  readonly id: string;
+  readonly name: string;
+}
+
+export class AIProviderRegistry extends ProviderRegistry<IAIProviderContract> {}