+++ src/providers/registry/OCRProviderRegistry.ts
@@ -1,0 +1,8 @@
+import { ProviderRegistry } from './ProviderRegistry';
+
+export interface IOCRProviderContract {
+  readonly id: string;
+  readonly name: string;
+}
+
+export class OCRProviderRegistry extends ProviderRegistry<IOCRProviderContract> {}