+++ src/providers/registry/SearchProviderRegistry.ts
@@ -1,0 +1,8 @@
+import { ProviderRegistry } from './ProviderRegistry';
+
+export interface ISearchProviderContract {
+  readonly id: string;
+  readonly name: string;
+}
+
+export class SearchProviderRegistry extends ProviderRegistry<ISearchProviderContract> {}