+++ src/providers/registry/SpeechProviderRegistry.ts
@@ -1,0 +1,8 @@
+import { ProviderRegistry } from './ProviderRegistry';
+
+export interface ISpeechProviderContract {
+  readonly id: string;
+  readonly name: string;
+}
+
+export class SpeechProviderRegistry extends ProviderRegistry<ISpeechProviderContract> {}