+++ src/providers/registry/StorageProviderRegistry.ts
@@ -1,0 +1,8 @@
+import { ProviderRegistry } from './ProviderRegistry';
+
+export interface IStorageProviderContract {
+  readonly id: string;
+  readonly name: string;
+}
+
+export class StorageProviderRegistry extends ProviderRegistry<IStorageProviderContract> {}